+++ src/EntityGraphQL/Extensions/StringExtensions.cs
@@ -15,5 +15,15 @@ namespace EntityGraphQL.Extensions
                 _ => input.First().ToString().ToUpper(CultureInfo.InvariantCulture) + input[1..]
             };
         }
+
+        public static string FirstCharToLower(this string input)
+        {
+            return input switch
+            {
+                null => throw new ArgumentNullException(nameof(input)),
+                "" => throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input)),
+                _ => input.First().ToString().ToLower(CultureInfo.InvariantCulture) + input[1..]
+            };
+        }
     }
 }