+++ src/EntityGraphQL/Schema/FieldExtensions/Sorting/SortExtension.cs
@@ -16,14 +16,12 @@ namespace EntityGraphQL.Schema.FieldExtensions
         private Type? methodType;
         private Func<string, string>? fieldNamer;
         private readonly Type? fieldSelectionType;
-        private readonly LambdaExpression? defaultSort;
-        private readonly SortDirection? defaultSortDirection;
+        private readonly List<ISort> defaultSorts;
 
-        public SortExtension(Type? fieldSelectionType, LambdaExpression? defaultSort, SortDirection? direction)
+        public SortExtension(Type? fieldSelectionType, params ISort[] defaultSorts)
         {
             this.fieldSelectionType = fieldSelectionType;
-            this.defaultSort = defaultSort;
-            this.defaultSortDirection = direction;
+            this.defaultSorts = defaultSorts?.ToList() ?? new List<ISort>();
         }
 
         public override void Configure(ISchemaProvider schema, IField field)
@@ -42,9 +40,10 @@ namespace EntityGraphQL.Schema.FieldExtensions
                 typeof(Queryable) : typeof(Enumerable);
 
             fieldNamer = schema.SchemaFieldNamer;
-            var sortInputName = $"{field.Name}SortInput".FirstCharToUpper();
+            var sortInputName = $"{field.FromType.Name}{field.Name.FirstCharToUpper()}SortInput".FirstCharToUpper();
             ISchemaType schemaSortType;
             var argSortType = MakeSortType(field);
+            // look type reuse type. Type is not recreated if it uses the same fields
             if (schema.HasType(argSortType))
                 schemaSortType = schema.GetSchemaType(argSortType, null);
             else
@@ -54,17 +53,24 @@ namespace EntityGraphQL.Schema.FieldExtensions
 
             var argType = typeof(SortInput<>).MakeGenericType(schemaSortType.TypeDotnet);
             var argInstance = Activator.CreateInstance(argType)!;
-            if (defaultSort != null)
+            if (defaultSorts.Any())
             {
-                var defaultSortValue = Activator.CreateInstance(schemaSortType.TypeDotnet)!;
-                // if the field is not there the default sort is not exposed in the API we the schema does not need to know about the default
-                var sortValueField = schemaSortType.TypeDotnet.GetField(((MemberExpression)defaultSort.Body).Member.Name);
-                if (sortValueField != null)
+                var defaultSortValues = Activator.CreateInstance(typeof(List<>).MakeGenericType(schemaSortType.TypeDotnet))!;
+
+                foreach (var defaultSort in defaultSorts)
                 {
-                    sortValueField.SetValue(defaultSortValue, defaultSortDirection ?? SortDirection.ASC);
-                    var defaultSortValues = Activator.CreateInstance(typeof(List<>).MakeGenericType(schemaSortType.TypeDotnet))!;
-                    ((IList)defaultSortValues).Add(defaultSortValue);
-                    argType.GetProperty("Sort")!.SetValue(argInstance, defaultSortValues);
+                    var defaultSortValue = Activator.CreateInstance(schemaSortType.TypeDotnet)!;
+                    // if the field is not there the default sort is not exposed in the API we the schema does not need to know about the default
+                    var fieldExp = defaultSort.SortExpression.Body;
+                    if (fieldExp.NodeType == ExpressionType.Convert)
+                        fieldExp = ((UnaryExpression)fieldExp).Operand;
+                    var sortValueField = schemaSortType.TypeDotnet.GetField(((MemberExpression)fieldExp).Member.Name);
+                    if (sortValueField != null)
+                    {
+                        sortValueField.SetValue(defaultSortValue, defaultSort.Direction);
+                        ((IList)defaultSortValues).Add(defaultSortValue);
+                        argType.GetProperty("Sort")!.SetValue(argInstance, defaultSortValues);
+                    }
                 }
             }
             field.AddArguments(argInstance);
@@ -138,16 +144,21 @@ namespace EntityGraphQL.Schema.FieldExtensions
                     sortMethod = "ThenBy";
                 }
             }
-            else if (defaultSort != null)
+            else if (defaultSorts.Any())
             {
-                var listParam = Expression.Parameter(listType!);
-                expression = Expression.Call(
-                        methodType!,
-                        defaultSortDirection == SortDirection.ASC ? "OrderBy" : "OrderByDescending",
-                        new Type[] { listType!, defaultSort.Body.Type },
-                        expression,
-                        parameterReplacer.Replace(defaultSort, defaultSort.Parameters.First(), listParam)
-                    );
+                var thenBy = false;
+                foreach (var defaultSort in defaultSorts)
+                {
+                    var listParam = Expression.Parameter(listType!);
+                    expression = Expression.Call(
+                            methodType!,
+                            defaultSort.Direction == SortDirection.ASC ? (thenBy ? "ThenBy" : "OrderBy") : (thenBy ? "ThenByDescending" : "OrderByDescending"),
+                            new Type[] { listType!, defaultSort.SortExpression.Body.Type },
+                            expression,
+                            parameterReplacer.Replace(defaultSort.SortExpression, defaultSort.SortExpression.Parameters.First(), listParam)
+                        );
+                    thenBy = true;
+                }
             }
             return expression;
         }